requirements.txt
streamlit
pandas
openpyxl
schedule
google-auth
google-auth-oauthlib
google-api-python-client
streamlit-javascript
